Patch nvshmem to avoid an uninitialized value passed to RoCE
NVSHMEM does not clear a stack allocated struct `ibv_ah_attr`
which is passed directly to the kernel RDMA functions. On
kernels with a patch to validate the value of static_rate
the create_ah() function will return EINVAL (sometimes).

When built in environments that automatically zero stack
values, the library will always initialize the device to
the default rate. Otherwise, it may result in the error

`Unable to create ah.` especially on RoCE devices.

Signed-off-by: Clayton Coleman <smarterclayton@gmail.com>